Document internal classifier and harness test interfaces (#517)
Add `///` documentation covering purpose, parameter meanings, return
interpretation, and mutation contracts to the argument-classifier
internals that lacked it:

- `ClassificationContext` (the accumulator/placeholder-set pairing and
  the claim-a-placeholder discipline).
- `classify_datatable` / `classify_docstring` /
  `classify_fixture_or_step` (matching rules, in-place attribute
  stripping, and every error condition).
- `extract_flag_attribute` / `parse_from_attribute` (explicitly noting
  the in-place `arg.attrs` mutation).

Expand the `rstest-bdd-harness/src/test_utils.rs` module header to
explain the panic-payload matching helpers, the `&str`/`String`
payload asymmetry they absorb, and when harness implementations should
reach for them.

The added documentation pushed `classify.rs` past the 400-line budget,
so the terminal classifier (`classify_fixture_or_step` with its
`#[from]` parsing and placeholder-match helpers) moves to a focused
`classify/fixture_or_step.rs` submodule.

The `cargo-bdd` output helpers listed in the issue received their
documentation (purpose plus leading-space/leading-newline contracts)
in the issue #513 consolidation (PR #529); re-documenting their
pre-consolidation shapes here would only manufacture merge conflicts.

`cargo doc` builds with no new warnings.
