+++ crates/rstest-bdd-macros/src/codegen/wrapper/args/classify.rs
@@ -12,6 +12,7 @@ use super::{Arg, ExtractedArgs, normalize_param_name};
 
 mod step_struct;
 
+pub(super) use fixture_or_step::classify_fixture_or_step;
 pub(super) use step_struct::{classify_step_struct, extract_step_struct_attribute};
 
 /// Diagnostic for a `datatable` parameter declared with an unsupported type.
@@ -36,8 +37,18 @@ const DATATABLE_AFTER_DOCSTRING_ERROR: &str =
 /// Diagnostic for combining `#[datatable]` with `#[from]`.
 const DATATABLE_WITH_FROM_ERROR: &str = "#[datatable] cannot be combined with #[from]";
 
+/// Mutable state threaded through the classification pipeline.
+///
+/// Bundles the accumulator of classified arguments (`extracted`) with the set
+/// of step-pattern placeholders not yet claimed by a parameter
+/// (`placeholders`). Classifiers that match a placeholder remove it from the
+/// set, so each placeholder binds at most one parameter; whatever remains
+/// after classification represents unbound placeholders.
 pub(super) struct ClassificationContext<'a> {
+    /// Accumulated classification results, appended to in place.
     pub(super) extracted: &'a mut ExtractedArgs,
+    /// Step-pattern placeholders still awaiting a matching parameter;
+    /// classifiers remove entries as they claim them.
     pub(super) placeholders: &'a mut HashSet<String>,
 }
 
@@ -99,6 +110,16 @@ fn should_classify_as_datatable(pat: &syn::Ident, ty: &syn::Type) -> bool {
     pat == "datatable" && (is_datatable(ty) || is_cached_table(ty))
 }
 
+/// Detect and strip a marker attribute (e.g. `#[datatable]`) from `arg`.
+///
+/// Mutates `arg.attrs` in place, removing every attribute whose path is
+/// `attr_name` so the generated wrapper does not re-emit it. Returns whether
+/// the attribute was present.
+///
+/// # Errors
+///
+/// Returns an error when the attribute carries arguments (the marker form
+/// takes none) or appears more than once on the same parameter.
 pub(super) fn extract_flag_attribute(arg: &mut syn::PatType, attr_name: &str) -> syn::Result<bool> {
     let mut found = false;
     let mut duplicate = false;
@@ -196,6 +217,30 @@ where
     }
 }
 
+/// Classify `arg` as the step’s `DataTable` parameter, if it matches.
+///
+/// A parameter matches when it is annotated `#[datatable]` or when it is the
+/// canonical `datatable: Vec<Vec<String>>` / `datatable: CachedTable` shape.
+/// On a match the argument is recorded in `st` (setting `st.datatable_idx`)
+/// and the `#[datatable]` marker attribute is stripped from `arg` in place
+/// via [`extract_flag_attribute`].
+///
+/// - `st` — classification accumulator, mutated on success.
+/// - `arg` — the parameter being classified; its attribute list is mutated
+///   in place even when validation subsequently fails.
+/// - `pat` / `ty` — the parameter's identifier and type, pre-extracted by
+///   the caller.
+///
+/// Returns `Ok(true)` when the parameter was claimed, `Ok(false)` to let the
+/// next classifier try.
+///
+/// # Errors
+///
+/// Returns an error when the parameter is named `datatable` with an
+/// unsupported type, when `#[datatable]` is applied to the reserved
+/// `docstring` parameter, when a `DataTable` was already classified, when the
+/// `DataTable` appears after a `DocString` (Gherkin ordering), or when
+/// `#[datatable]` is combined with `#[from]`.
 pub(super) fn classify_datatable(
     st: &mut ExtractedArgs,
     arg: &mut syn::PatType,
@@ -244,6 +289,20 @@ fn is_docstring_canonical(pat: &syn::Ident, ty: &syn::Type) -> bool {
     pat == "docstring" && is_string(ty)
 }
 
+/// Classify `arg` as the step’s `DocString` parameter, if it matches.
+///
+/// A parameter matches when it is the canonical `docstring: String` shape.
+/// On a match the argument is recorded in `st` (setting `st.docstring_idx`).
+/// Unlike [`classify_datatable`] there is no marker attribute, so `arg` is
+/// not mutated.
+///
+/// Returns `Ok(true)` when the parameter was claimed, `Ok(false)` to let the
+/// next classifier try.
+///
+/// # Errors
+///
+/// Returns an error when a parameter named `docstring` has a type other than
+/// `String`, or when a `DocString` parameter was already classified.
 pub(super) fn classify_docstring(
     st: &mut ExtractedArgs,
     arg: &mut syn::PatType,
@@ -271,101 +330,7 @@ pub(super) fn classify_docstring(
     st.docstring_idx = Some(idx);
     Ok(true)
 }
-
-fn parse_from_attribute(arg: &mut syn::PatType) -> syn::Result<Option<syn::Ident>> {
-    let mut from_name = None;
-    let mut from_attr_err = None;
-    arg.attrs.retain(|a| {
-        if !a.path().is_ident("from") {
-            return true;
-        }
-        if from_attr_err.is_some() {
-            return false;
-        }
-        match &a.meta {
-            syn::Meta::Path(_) => {}
-            syn::Meta::List(_) => match a.parse_args::<syn::Ident>() {
-                Ok(parsed) => from_name = Some(parsed),
-                Err(err) => from_attr_err = Some(err),
-            },
-            syn::Meta::NameValue(_) => {
-                from_attr_err = Some(syn::Error::new_spanned(
-                    a,
-                    "#[from] expects an identifier or no arguments",
-                ));
-            }
-        }
-        false
-    });
-    if let Some(err) = from_attr_err {
-        return Err(err);
-    }
-    Ok(from_name)
-}
-
-fn validate_no_step_struct_conflict(
-    ctx: &ClassificationContext,
-    target_name: &str,
-    pat: &syn::Ident,
-) -> syn::Result<()> {
-    if ctx.extracted.step_struct_idx.is_some()
-        && ctx.extracted.blocked_placeholders.contains(target_name)
-    {
-        Err(syn::Error::new(
-            pat.span(),
-            "#[step_args] cannot be combined with named step arguments",
-        ))
-    } else {
-        Ok(())
-    }
-}
-
-fn classify_by_placeholder_match(
-    ctx: &mut ClassificationContext,
-    from_name: Option<syn::Ident>,
-    pat: syn::Ident,
-    ty: syn::Type,
-) -> syn::Result<()> {
-    let target = from_name.clone().unwrap_or_else(|| pat.clone());
-    let target_name = target.to_string();
-    let normalized = normalize_param_name(&target_name);
-    if ctx.placeholders.remove(normalized) {
-        validate_no_step_struct_conflict(ctx, normalized, &pat)?;
-        ctx.extracted.push(Arg::Step { pat, ty });
-        Ok(())
-    } else {
-        validate_no_step_struct_conflict(ctx, &target_name, &pat)?;
-        let name = if let Some(name) = from_name {
-            name
-        } else if normalized == target_name {
-            pat.clone()
-        } else {
-            let mut name = syn::parse_str::<syn::Ident>(normalized).map_err(|_| {
-                syn::Error::new(
-                    pat.span(),
-                    format!(
-                        "normalized fixture name `{normalized}` is not a valid identifier; use #[from(...)] to specify the fixture name explicitly"
-                    ),
-                )
-            })?;
-            name.set_span(pat.span());
-            name
-        };
-        ctx.extracted.push(Arg::Fixture { pat, name, ty });
-        Ok(())
-    }
-}
-
-pub(super) fn classify_fixture_or_step(
-    ctx: &mut ClassificationContext,
-    arg: &mut syn::PatType,
-    pat: syn::Ident,
-    ty: syn::Type,
-) -> syn::Result<bool> {
-    let from_name = parse_from_attribute(arg)?;
-    classify_by_placeholder_match(ctx, from_name, pat, ty)?;
-    Ok(true)
-}
-
 #[cfg(test)]
 mod tests;
+
+mod fixture_or_step;
